update page structure
